cuda: compatibility layer for Jetson Nano (CUDA 10.2 / sm_53)
Emulate bf16 via a compat header (no cuda_bf16.h pre-CUDA-11), guard the bf16
cuBLAS GEMM paths (no CUDA_R_16BF) so bf16 routes to fp32, disable CUDA graphs
on nvcc < 11, and guard the CUDA 11 cooperative_groups/reduce.h include. bf16
dequant restored to nv_bfloat16 so the shim handles it correctly. Adds
docs/jetson-nano.md. All changes guarded on CUDART_VERSION / nvcc version, so
CUDA 11+ builds are unaffected.

+++ ggml/src/ggml-cuda/convert.cu
@@ -761,7 +761,7 @@ to_fp16_cuda_t ggml_get_to_fp16_cuda(ggml_type type) {
         case GGML_TYPE_F32:
             return convert_unary_cont_cuda<float>;
         case GGML_TYPE_BF16:
-            return convert_unary_cont_cuda<half>;
+            return convert_unary_cont_cuda<nv_bfloat16>;
         default:
             return nullptr;
     }
@@ -816,7 +816,7 @@ to_fp32_cuda_t ggml_get_to_fp32_cuda(ggml_type type) {
         case GGML_TYPE_F16:
             return convert_unary_cont_cuda<half>;
         case GGML_TYPE_BF16:
-            return convert_unary_cont_cuda<half>;
+            return convert_unary_cont_cuda<nv_bfloat16>;
         default:
             return nullptr;
     }
@@ -839,7 +839,7 @@ to_fp16_nc_cuda_t ggml_get_to_fp16_nc_cuda(ggml_type type) {
         case GGML_TYPE_Q8_0:
             return dequantize_block_cuda<QK8_0, QR8_0, dequantize_q8_0>;
         case GGML_TYPE_BF16:
-            return convert_unary_cuda<half>;
+            return convert_unary_cuda<nv_bfloat16>;
         default:
             return nullptr;
     }
